Make it easier to feature-gate entire bench program
